feat(vmm): add tap device creation
This adds the automatic creation of a tap network interface on the host
when a VMM is started. Those network interfaces are named `tapletX`.

Two new CLI arguments were added to configure this interface, `--network-host-ip`
and `--network-host-netmask`, which configures the IP address of the tap
interface on the host side.

Code was taken and adapted from the [cloud-hypervisor project](https://github.com/cloud-hypervisor/cloud-hypervisor/tree/main/net_util/src).

Signed-off-by: Kuruyia <github@kuruyia.net>