feat(vmm): exit the VMM on kbd reset
Currently, when killing the `init` process, the VMM hangs because it is
unable to handle the kernel trying to reboot the machine.

On an x86 PC platform, it is possible to reset the CPU by [sending byte
`0xFE` to the keyboard controller command register](https://wiki.osdev.org/%228042%22_PS/2_Controller#CPU_Reset).
This is actually part of the emergency reboot procedure of [the Linux kernel](https://elixir.bootlin.com/linux/v6.8.2/source/arch/x86/kernel/reboot.c#L627).

This adds a bit of code to handle such a write operation by exiting the
VMM process.

Signed-off-by: Kuruyia <github@kuruyia.net>